Create LICENSE
CC BY 4.0 LICENSE
The Creative Commons Attribution 4.0 International (CC BY 4.0) license is an open license that allows others to use, share, adapt, and build upon your work, even for commercial purposes, as long as attribution is given to the original creator. This license is specifically designed to promote openness and broad sharing, with the flexibility for reusers to remix or transform the material to fit different needs.